Move code out of definitions
This moves builtin-creation out of mathics.builtin and core.definitions
and into a new module.

I thing this is the last step before we actually move builtin creation
of the Definitions initialization and into some one-time code.
