+++ mathics/builtin/__init__.py
@@ -31,6 +31,9 @@ __py_files__ = [
     for f in glob.glob(osp.join(osp.dirname(__file__), "[a-z]*.py"))
 ]
 
+system_builtins_dict = {}
+builtins_by_module = {}
+
 
 def add_builtins(new_builtins):
     for var_name, builtin in new_builtins:
@@ -48,7 +51,7 @@ def add_builtins(new_builtins):
             builtins_precedence[name] = builtin.precedence
         if isinstance(builtin, PatternObject):
             pattern_objects[name] = builtin.__class__
-    _builtins.update(dict(new_builtins))
+    system_builtins_dict.update(dict(new_builtins))
 
 
 def builtins_dict():
@@ -57,26 +60,6 @@ def builtins_dict():
         for modname, builtins in builtins_by_module.items()
         for builtin in builtins
     }
-
-
-def contribute(definitions):
-    # let MakeBoxes contribute first
-    _builtins["System`MakeBoxes"].contribute(definitions)
-    for name, item in _builtins.items():
-        if name != "System`MakeBoxes":
-            item.contribute(definitions)
-
-    from mathics.core.definitions import Definition
-    from mathics.core.expression import ensure_context
-    from mathics.core.parser import all_operator_names
-
-    # All builtins are loaded. Create dummy builtin definitions for
-    # any remaining operators that don't have them. This allows
-    # operators like \[Cup] to behave correctly.
-    for operator in all_operator_names:
-        if not definitions.have_definition(ensure_context(operator)):
-            op = ensure_context(operator)
-            definitions.builtin[op] = Definition(name=op)
 
 
 def import_builtins(module_names: List[str], submodule_name=None) -> None:
@@ -162,7 +145,6 @@ modules = []
 import_builtins(module_names)
 
 _builtins_list = []
-builtins_by_module = {}
 
 disable_file_module_names = (
     [] if ENABLE_FILES_MODULE else ["files_io.files", "files_io.importexport"]
@@ -257,9 +239,6 @@ sympy_to_mathics = {}
 builtins_precedence = {}
 
 new_builtins = _builtins_list
-
-# FIXME: some magic is going on here..
-_builtins = {}
 
 add_builtins(new_builtins)
