+++ mathics/session.py
@@ -12,7 +12,7 @@ In particular we provide:
 import os.path as osp
 from typing import Optional
 
-from mathics.core.definitions import autoload_files
+from mathics.builtin.system_init import autoload_files
 from mathics.core.parser import parse, MathicsSingleLineFeeder
 from mathics.core.definitions import Definitions
 from mathics.core.evaluation import Evaluation
